Fix submodules support
